--- packages/next/vitals/index.ts
@@ -1,1 +1,0 @@
-export { useExperimentalWebVitalsReport } from './vitals'
